Create Power test of CMH test
While applications of Cochran-Mantel-Haenszel (CMH) test commonly use sequencing read count as an estimation of allele count, this will only be accurate when the haploid size of the sample is much greater than the depth of coverage (so that most reads arise from uniquely sampled haplotypes). As the pool size used in the present study (diploid size of 2N=10) is often smaller than the sequencing depth of coverage, read counts were converted into allele frequencies and then multiplied by the ploidy to yield an approximate count for each pool that represents the real replication level. We conducted a simulation to compare the false positive rate from the CMH test using uncorrected or corrected allele counts across a range of number of demes, individuals, and depths.